== → = for proper compliance with bash & sh.
Add KTX_FEATURE_TOOLS to scripts.

+++ ci_scripts/install_macos.sh
@@ -7,16 +7,26 @@
 # Exit if any command fails.
 set -e
 
+# Set parameters from command-line arguments, if any.
+for i in $@; do
+  eval $i
+done
+
+FEATURE_LOADTESTS=${FEATURE_LOADTESTS:-ON}
+
 git lfs install
 git lfs version
-brew link sdl2
 gem install xcpretty
 
-# Current directory is .../build/{KhronosGroup,msc-}/KTX-Software. cd to 'build'.
-pushd ../..
-wget -O vulkansdk-macos-$VULKAN_SDK_VER.dmg https://sdk.lunarg.com/sdk/download/$VULKAN_SDK_VER/mac/vulkansdk-macos-$VULKAN_SDK_VER.dmg?Human=true
-hdiutil attach vulkansdk-macos-$VULKAN_SDK_VER.dmg
-sudo /Volumes/vulkansdk-macos-$VULKAN_SDK_VER/InstallVulkan.app/Contents/macOS/InstallVulkan --root "$VULKAN_INSTALL_DIR" --accept-licenses --default-answer --confirm-command install
-hdiutil detach /Volumes/vulkansdk-macos-$VULKAN_SDK_VER
-rm vulkansdk-macos-$VULKAN_SDK_VER.dmg
-popd
+if [ "$FEATURE_LOADTESTS" = "ON" ]; then
+  brew link sdl2
+
+  # Current dir. is .../build/{KhronosGroup,msc-}/KTX-Software. cd to 'build'.
+  pushd ../..
+  wget -O vulkansdk-macos-$VULKAN_SDK_VER.dmg https://sdk.lunarg.com/sdk/download/$VULKAN_SDK_VER/mac/vulkansdk-macos-$VULKAN_SDK_VER.dmg?Human=true
+  hdiutil attach vulkansdk-macos-$VULKAN_SDK_VER.dmg
+  sudo /Volumes/vulkansdk-macos-$VULKAN_SDK_VER/InstallVulkan.app/Contents/macOS/InstallVulkan --root "$VULKAN_INSTALL_DIR" --accept-licenses --default-answer --confirm-command install
+  hdiutil detach /Volumes/vulkansdk-macos-$VULKAN_SDK_VER
+  rm vulkansdk-macos-$VULKAN_SDK_VER.dmg
+  popd
+fi
